refactor: path -> map 패키지로 이동
