+++ src/main/java/nextstep/subway/map/exception/NotConnectedSourceAndTargetException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.path.exception;
+package nextstep.subway.map.exception;
 
 public class NotConnectedSourceAndTargetException extends RuntimeException {
     public NotConnectedSourceAndTargetException() {
