+++ src/main/java/nextstep/subway/map/ui/PathController.java
@@ -1,8 +1,8 @@
-package nextstep.subway.path.ui;
+package nextstep.subway.map.ui;
 
-import nextstep.subway.path.application.PathService;
-import nextstep.subway.path.dto.PathRequest;
-import nextstep.subway.path.dto.PathResponse;
+import nextstep.subway.map.application.PathService;
+import nextstep.subway.map.dto.PathRequest;
+import nextstep.subway.map.dto.PathResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
